Added Gato and AI classes, added code to classes.
Copied the game's code from file 'BallmerPeak' to one of the new clases.
Tried the game with a temoral UI. It doesn't work but i'll get to it later.
The issue appear to be in the UI and also the game let you repeat the places you
make a mark.